revert some'
