--- tests/config.rs
@@ -1,1 +1,0 @@
-../src/config.rs
