--- tests/errors.rs
@@ -1,1 +1,0 @@
-../src/errors.rs
