--- tests/tls.rs
@@ -1,1 +1,0 @@
-../src/tls.rs
